Moved lots of code out of main.c/cleanup.
