+++ include/arm11/open_agb_firm.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "error_codes.h"
+
+
+
+Result oafInitAndRun(void);
+void oafUpdate(void);
+void oafFinish(void);
